Time: 302 ms (52.39%), Space: 84.3 MB (54.09%) - LeetHub

+++ 2226-maximum-candies-allocated-to-k-children/2226-maximum-candies-allocated-to-k-children.cpp
@@ -21,6 +21,6 @@ public:
                 end=mid-1;
             
         }
-        return ans;
+        return (int)ans;
     }
 };
